Uppfæra glærupakka 4 fyrir 2018. Endursníða öll sýnidæmi, fækka STL gagnagrindum sem farið er yfir

+++ T2/Code/w3/recursivesum.cpp
@@ -1,16 +1,16 @@
 #include <iostream>
 using namespace std;
 
-long recsum(long n) { // Útreikningur á sum_{i=0}^n i
+int recsum(int n) {  // Útreikningur á sum_{i=0}^n i
     if (n <= 0) {
         return 0;
     } else {
-        return n + recsum(n-1); // Útreikningum ekki lokið
+        return n + recsum(n - 1);  // Útreikningum ekki lokið
     }
 }
 
 int main() {
-    long i;
+    int i;
     cin >> i;
     cout << recsum(i) << endl;
     return 0;
